Singleton Initialization
Refactored a lot of logic to try and better organize the various
singletons and to streamline/simplify the initialization of objects.
Unlike with spring's dependency injection which manages the singletons
behind its veil, a context holder for singles are needed for the system
to minimize code and to make it more streamlined to develop within this
framework.