Added Logging Wrapper
Makes it a little bit easier to use the java.util.Logger system and
still retain some of the nice benefits that things like log4j / logback
offer.

It will autodetect if the last arguement is a Throwable and if so auto
set it as an causing exception. Also allows efficient var args style
parameterization of messages which only are processed if the log level
is appropriately met.