Basic Injection Support
Added a very trivial and lightweight support for handling dependency
injection based on the java JSR standard. The implementation in this
version houses a manager class called InjectionHolder which holds the
references to all the objects and there keys which are injectable into
other classes. Also has a .inject method to take any arbitrary object
and locate all the @Inject annotations and associating the references
that are registered with the InjectionHolder.

It becomes possible to later externalize the bean creation and configs
to an external xml file the same way spring does for its IoC. This
would allow the clean and familiar usage of the system without all the
overhead and baggage that comes along with spring.