Improved JSON response without customization
Now collections and maps with complex objects in them will by default
work a bit better. Before to get them to output correct, the Response
interface had to be implemented on a DTO but now an implicit DTO
response is generated using all the getters of the objects in the list
or map.